problems/src/test/java/_1001_1050/_1004_Max_Consecutive_Ones_III/SolutionITest.java
package _1001_1050._1004_Max_Consecutive_Ones_III;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 10/18/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final int A[] = {1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0}, K = 2;
    assertEquals(6, solution.longestOnes(A, K));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final int A[] = {0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1}, K = 3;
    assertEquals(10, solution.longestOnes(A, K));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}